requirements.txt
streamlit
openai
faiss-cpu
beautifulsoup4
requests
python-dotenv
